Data.Bytes.Class.Conversion

Description

Provides typeclasses for convert between byte sizes.

Since: 0.1

High-level

class Conversion a where

This class allows one to transform a bytes type to any Size. For types with existentially quantified Size (e.g. SomeSize, SomeNetSize), this will "undo" the existential quantification.

Since: 0.1

Associated Types

type Converted (t :: Size) a = (r :: Type) | r -> t

Since: 0.1

Methods

convert_ :: forall (t :: Size). SingI t => a -> Converted t a

convert_ @_ @t x converts x to size t.

Examples

Expand

>>> let bytes = MkBytes 50_000 :: Bytes 'M Int
>>> let gBytes = convert_ @_ @G bytes
>>> :type gBytes
gBytes :: Bytes G Int
>>> gBytes
MkBytes 50

>>> let bytes = hideSize (MkBytes 0.2 :: Bytes 'T Float)
>>> let mBytes = convert_ @_ @M bytes
>>> :type mBytes
mBytes :: Bytes M Float
>>> mBytes
MkBytes 200000.0

Since: 0.1

Low-level

convertSize :: (FromInteger n, MGroup n) => Size -> Size -> n -> n

Low level function for converting a numeric literal between byte sizes. convert b1 b2 converts from b1 to b2, e.g. convert G K = \n -> n * 1_000_000. The higher level byte types and functions should be preferred (e.g. Conversion, Data.Bytes.Class.Normalize), but this is here when it is needed.

Since: 0.1

convertWitness :: forall (s :: Size) n. (FromInteger n, MGroup n, SingI s) => Size -> n -> n

Low level function for converting a numeric literal from the inferred SingI to the parameter Size. For instance,

>>> convertWitness @K M 5_000
5

This is slightly more principled than convertSize, but the higher level byte types and functions should still be preferred (e.g. Bytes, Normalize).

Since: 0.1