Data.Bytes.Formatting

Description

Provides high level formatters for different byte types. This module contains the functions and types for typical usage. Additional functionality (e.g. optics, type classes/families for defining custom formatters) can be found in:

Data.Bytes.Formatting.Base

Data.Bytes.Formatting.Size

Data.Bytes.Formatting.Direction

Since: 0.1

High-level functions

formatSized

Arguments

:: (Formatter (BaseFormatter (Raw a)), PrintfArg (Raw a), RawNumeric a, Sized a)

=> BaseFormatter (Raw a)
Formatter to use on the underlying value i.e. IntegralFormatter or FloatingFormatter.

-> SizedFormatter
Formatter to use on the size units.

-> a
Value to be formatted.

-> Text
Result.

Formats a value with Size units. Can be used with all byte types in this package.

Examples

Expand

>>> import Data.Default (Default (..))
>>> import Data.Bytes (Bytes (..), Size (..), hideSize)
>>> let bf1 = MkIntegralFormatter
>>> let b1 = MkBytes @M @Int 7
>>> formatSized bf1 def b1
"7 mb"

>>> formatSized bf1 sizedFormatterUnix b1
"7M"

>>> let bf2 = MkFloatingFormatter (Just 2)
>>> let b2 = hideSize $ MkBytes @G @Float 20.248
>>> formatSized bf2 sizedFormatterVerbose b2
"20.25 gigabytes"

Since: 0.1

formatSizedDirected

Arguments

:: (Directed a, Formatter (BaseFormatter (Raw a)), PrintfArg (Raw a), RawNumeric a, Sized a)

=> BaseFormatter (Raw a)
Formatter to use on the underlying value i.e. IntegralFormatter or FloatingFormatter.

-> SizedFormatter
Formatter to use on the size units.

-> DirectedFormatter
Formatter to use on the direction units.

-> a
Value to be formatted.

-> Text
Result.

Formats a value with Size and Direction units. Can only be used when both units are present.

Examples

Expand

>>> import Data.Default (Default (..))
>>> import Data.Bytes.Network (Direction (..), NetBytes (..), Size (..))
>>> import Data.Bytes.Size (Sized (..))
>>> let bf1 = MkIntegralFormatter
>>> let b1 = MkNetBytesP @Up @M @Int 7
>>> formatSizedDirected bf1 def def b1
"7 mb up"

>>> formatSizedDirected bf1 sizedFormatterUnix directedFormatterUnix b1
"7M U"

>>> let bf2 = MkFloatingFormatter (Just 2)
>>> let b2 = hideSize $ MkNetBytesP @Down @G @Float 20.248
>>> formatSizedDirected bf2 sizedFormatterVerbose directedFormatterVerbose b2
"20.25 gigabytes down"

Since: 0.1

Types

Base

data CaseFormat

Case formatting.

Since: 0.1

Constructors

CaseFormatLower
Since: 0.1

CaseFormatTitle
Since: 0.1

CaseFormatUpper
Since: 0.1

data IntegralFormatter

Formatter for integral types.

Since: 0.1

Constructors

MkIntegralFormatter

newtype FloatingFormatter

Formatter for floating types. Takes an optional param for rounding digits.

Since: 0.1

Constructors

MkFloatingFormatter (Maybe Word8)

Size

data SizeFormat

Determines how to format the size units.

Examples

Expand

>>> def @SizeFormat
SizeFormatMedium

Since: 0.1

Constructors

SizeFormatShort
Since: 0.1

SizeFormatMedium
Since: 0.1

SizeFormatLong
Since: 0.1

data SizedFormatter

Formatting size units.

Examples

Expand

>>> def @SizedFormatter
MkSizedFormatter {caseFormat = CaseFormatLower, leadingSpace = True, sizeFormat = SizeFormatMedium}

Since: 0.1

Constructors

MkSizedFormatter CaseFormat Bool SizeFormat

sizedFormatterUnix :: SizedFormatter

Unix-style formatter:

Upper case.

No space between value and unit.

Single letter unit.

For example: 7G.

Since: 0.1

sizedFormatterNatural :: SizedFormatter

Natural-style formatter:

Lower case.

Space between value and unit.

Two letter units.

For example: 7 gb.

Since: 0.1

sizedFormatterVerbose :: SizedFormatter

Verbose formatter:

Lower case.

Space between value and unit.

Full spelling.

For example: 7 gigabytes.

Since: 0.1

Direction

data DirectionFormat

Determines how to format the direction units.

Examples

Expand

>>> def @DirectionFormat
DirectionFormatLong

Since: 0.1

Constructors

DirectionFormatShort
Since: 0.1

DirectionFormatLong
Since: 0.1

data DirectedFormatter

Formats bytes over floating types.

Examples

Expand

>>> def @DirectedFormatter
MkDirectedFormatter {caseFormat = CaseFormatLower, directionFormat = DirectionFormatLong}

Since: 0.1

Constructors

MkDirectedFormatter CaseFormat DirectionFormat

directedFormatterUnix :: DirectedFormatter

"unix"-style formatter:

Upper case.

Short spelling.

For example: 7 U.

Since: 0.1

directedFormatterVerbose :: DirectedFormatter

Verbose formatter:

Lower case.

Full spelling.

For example: 7 down.

Since: 0.1

Reexports

class Default a where

A class for types with a default value.

Minimal complete definition

Nothing

Methods

def :: a

The default value for this type.

default def :: (Generic a, GDefault (Rep a)) => a