Data.Bytes.Network

Description

This module serves as an alternative entry point to Data.Bytes, for when tracking uploaded vs. downloaded bytes is necessary. It provides the types and operations for typical usage and is usually the only import required. The core concept is:

Wrapping a numeric value representing bytes in a new type.

Attaching phantom labels representing the units (e.g. B, K, M, ...).

Attaching phantom labels representing the direction (i.e. Down, Up).

This prevents mistakes, such as adding two different byte sizes/directions, or converting between sizes incorrectly.

Since: 0.1

Introduction

The main idea is to attach phantom labels to the numeric bytes, so we can track the size and direction units. This allows us to safely manipulate byte values without mixing up units, performing incorrect conversions, etc.

The core types are a newtype wrapper NetBytes, the Size units, and the Direction units:

newtype NetBytes (d :: Direction) (s :: Size) n

Wrapper around the Bytes type that adds the Direction tag.

Examples

Expand

>>> MkNetBytesP @Up @M 1000
MkNetBytes (MkBytes 1000)

Since: 0.1

Constructors

MkNetBytes (Bytes s n)

Bundled Patterns

pattern MkNetBytesP :: forall d s n. n -> NetBytes d s n
Pattern for de/constructing NetBytes.

Since: 0.1

data Size

Byte units.

Since: 0.1

Constructors

B
Bytes

Since: 0.1

K
Kilobytes

Since: 0.1

M
Megabytes

Since: 0.1

G
Gigabytes

Since: 0.1

T
Terabytes

Since: 0.1

P
Petabytes

Since: 0.1

E
Exabytes

Since: 0.1

Z
Zettabytes

Since: 0.1

Y
Yottabytes

Since: 0.1

data Direction

Tags for differentiating downloaded vs. uploaded bytes.

Since: 0.1

Constructors

Down
Since: 0.1

Up
Since: 0.1

Basic Usage

Construction

There are several ways to construct a NetBytes type.

FromInteger

>>> fromZ 80 :: NetBytes Up M Int
MkNetBytes (MkBytes 80)

Directly

>>> import Data.Bytes (Bytes (MkBytes))
>>> MkNetBytes (MkBytes 80) :: NetBytes Down M Int
MkNetBytes (MkBytes 80)

>>> -- using the @MkNetBytesP :: n -> NetBytes d s n@ pattern synonym
>>> MkNetBytesP 80 :: NetBytes Up K Int
MkNetBytes (MkBytes 80)

Optics (optics-core)

>>> -- 1. Using _MkNetBytes to construct a NetBytes from a Bytes
>>> import Optics.Core (review, (%))
>>> import Data.Bytes (_MkBytes)
>>> (review _MkNetBytes (MkBytes 80)) :: NetBytes Up G Int -- Bytes -> NetBytes
MkNetBytes (MkBytes 80)

>>> -- 2. Using _MkNetBytes and _MkBytes to construct a NetBytes from a
>>> --    numeric value.
>>> (review (_MkNetBytes % _MkBytes) 70) :: NetBytes Up G Int -- n -> NetBytes
MkNetBytes (MkBytes 70)

>>> -- 3. Using #unNetBytes to construct a NetBytes from a numeric value.
>>> (review #unNetBytes 70) :: NetBytes Up G Int -- n -> NetBytes
MkNetBytes (MkBytes 70)

Unknown Size

We sometimes have to deal with unknown sizes at runtime, which presents a problem. We handle this with the SomeNetSize type, which existentially quantifies the Size:

data SomeNetSize (d :: Direction) n

Wrapper for NetBytes, existentially quantifying the size. This is useful when a function does not know a priori what size it should return e.g.

>>> :{
  getUpTraffic :: IO (SomeNetSize Up Float)
  getUpTraffic = do
    -- getUpTrafficRaw :: IO (Float, String)
    (bytes, units) <- getUpTrafficRaw
    pure $ case units of
      "B" -> hideSize $ MkNetBytesP @Up @B bytes
      "K" -> hideSize $ MkNetBytesP @Up @K bytes
      _ -> error "todo"
:}

We define an equivalence relation on SomeNetSize that takes units into account. For instance,

>>> hideSize (MkNetBytesP @Up @K 1000) == hideSize (MkNetBytesP @Up @M 1)
True

Because we expose the underlying NetBytes in several ways (e.g. Show, the SSize witness), this is technically unlawful for equality as it breaks the extensionality law:

\[ x = y \implies f(x) = f(y). \]

Since: 0.1

Sized

Fortunately, we do not have to directly use the constructor or singletons. We can instead use the Sized class.

class Sized a where

Types that have a size.

Since: 0.1

Associated Types

type HideSize a

Type used to hide the size.

Since: 0.1

Methods

sizeOf :: a -> Size

Retrieves the size.

Examples

Expand

>>> import Data.Bytes (Bytes (..))
>>> sizeOf (MkBytes @G 7)
G

>>> sizeOf (hideSize $ MkBytes @M 7)
M

>>> import Data.Bytes.Network (NetBytes (..), Direction (..))
>>> sizeOf (hideSize $ MkNetBytesP @Up @M 7)
M

Since: 0.1

hideSize :: a -> HideSize a

Hides the size.

Examples

Expand

>>> import Data.Bytes (Bytes (..))
>>> hideSize (MkBytes @G 7)
MkSomeSize SG (MkBytes 7)

Optics

Once again, we can use optics for this.

>>> import Optics.Core (review)
>>> review _MkSomeNetSize (MkNetBytesP 70 :: NetBytes Down G Int)
MkSomeNetSize SG (MkNetBytes (MkBytes 70))

Unknown Direction

Like sizes, we can also handle unknown directions at runtime. The types involved here are SomeNetDir (hiding direction only) and SomeNet (hiding both direction and size).

data SomeNetDir (s :: Size) n

Wrapper for NetBytes, existentially quantifying the direction. This is useful when a function does not know a priori what direction it should return e.g.

>>> :{
  getMaxTraffic :: IO (SomeNetDir K Double)
  getMaxTraffic = do
    -- getMaxTraffickRaw :: IO (Double, String)
    (bytes, direction) <- getMaxTraffickRaw
    pure $ case direction of
      "down" -> hideDirection $ MkNetBytesP @Down bytes
      "up" -> hideDirection $ MkNetBytesP @Up bytes
      _ -> error "bad direction"
:}

We deliberately do not provide instances for SomeX classes that could be used to combine arbitrary SomeNetDirs (e.g. Applicative, AGroup), as that would defeat the purpose of enforcing the distinction between upload and downloaded bytes.

Equality is determined by the usual equivalence class -- that takes units into account -- and by considering the direction.

>>> let x = MkNetBytesP @Up @K 1000 :: NetBytes Up K Int
>>> let y = MkNetBytesP @Down @K 1000 :: NetBytes Down K Int
>>> hideDirection x == hideDirection x
True
>>> hideDirection x == hideDirection y
False

Notice no Ord instance is provided, as we provide no ordering for Direction.

Since: 0.1

data SomeNet n

Wrapper for NetBytes, existentially quantifying the size and direction. This is useful when a function does not know a priori what size or direction it should return e.g.

>>> :{
  getMaxTraffic :: IO (SomeNet Double)
  getMaxTraffic = do
    -- getMaxTrafficNetRaw :: IO (Double, String, String)
    (bytes, direction, size) <- getMaxTrafficNetRaw
    pure $ case (direction, size) of
      ("Down", "K") -> hideDirection $ hideSize $ MkNetBytesP @Down @K bytes
      ("Up", "M") -> hideDirection $ hideSize $ MkNetBytesP @Up @M bytes
      _ -> error "todo"
:}

SomeNet carries along SDirection and SSize runtime witnesses for recovering the Direction and Size, respectively.

This instance uses the same equivalence relation from SomeNetSize w.r.t the size, and also includes an equality check on the direction. Thus we have, for instance,

>>> let x = MkNetBytesP 1_000 :: NetBytes Up K Int
>>> let y = MkNetBytesP 1 :: NetBytes Up M Int
>>> let z = MkNetBytesP 1_000 :: NetBytes Down K Int
>>> hideDirection (hideSize x) == hideDirection (hideSize y)
True

>>> hideDirection (hideSize x) == hideDirection (hideSize z)
False

Since: 0.1

Analogous to the Sized class, we have Directed.

class Directed a where

Types that have a direction.

Since: 0.1

Associated Types

type HideDirection a

Type used to hide the size.

Since: 0.1

Methods

directionOf :: a -> Direction

Retrieve the direction.

Examples

Expand

>>> import Data.Bytes.Network
>>> directionOf (MkNetBytesP @Up @G 7)
Up

>>> directionOf (hideSize $ hideDirection $ MkNetBytesP @Down @M 100)
Down

Since: 0.1

hideDirection :: a -> HideDirection a

Hides the direction.

Examples

Expand

>>> import Data.Bytes.Network (NetBytes (..), Size (..))
>>> hideDirection (MkNetBytesP @Up @G 7)
MkSomeNetDir SUp (MkNetBytes (MkBytes 7))

Since: 0.1

Once again, we can use optics for this.

>>> import Optics.Core (review)
>>> let x = MkNetBytesP 70 :: NetBytes Down G Int
>>> review _MkSomeNetDir x
MkSomeNetDir SDown (MkNetBytes (MkBytes 70))

>>> review _MkSomeNet x
MkSomeNet SDown SG (MkNetBytes (MkBytes 70))

Elimination

RawNumeric

We provide the RawNumeric class for conveniently unwrapping a type to the underlying numeric value.

class RawNumeric a where

Abstracts "wrapper" types for generically retrieving a raw numeric value.

Since: 0.1

Associated Types

type Raw a

Since: 0.1

Methods

toRaw :: a -> Raw a

Retrieves the underlying value.

Examples

Expand

>>> import Data.Bytes (Bytes (..), Size (..), Sized (..))
>>> toRaw (MkBytes @G 7)
7

>>> toRaw (hideSize $ MkBytes @M 400)
400

>>> import Data.Bytes.Network (Direction (..), NetBytes (..))
>>> toRaw (MkNetBytesP @Up @G 7)
7

>>> toRaw (hideSize $ MkNetBytesP @Up @G 7)
7

Since: 0.1

HasField

We can use HasField for this too.

>>> -- {-# LANGUAGE OverloadedRecordDot #-}
>>> let x1 = MkNetBytesP 7 :: NetBytes Up G Int
>>> x1.unNetBytes
7

>>> let x2 = hideSize x1 :: SomeNetSize Up Int
>>> x2.unSomeNetSize
7

>>> let x3 = (hideDirection x1) :: SomeNetDir G Int
>>> x3.unSomeNetDir
7

>>> let x4 = (hideDirection x2) :: SomeNet Int
>>> x4.unSomeNet
7

Optics

Optics are another option. The underscore-prefixed optics unwrap one level at a time, since we can freely compose them.

>>> import Optics.Core (view, (%))
>>> import Data.Bytes (_MkBytes)
>>> let x1 = MkNetBytesP 7 :: NetBytes Up G Int
>>> view (_MkNetBytes % _MkBytes) x1
7

>>> -- notice we have to convert the numeric value since the requested
>>> -- return type ('M') differs from the original ('G')
>>> let x2 = hideSize x1 :: SomeNetSize Up Int
>>> (view _MkSomeNetSize x2) :: NetBytes Up M Int
MkNetBytes (MkBytes 7000)

>>> view (_MkSomeNetSize % (_MkNetBytes @M) % _MkBytes) x2
7000

The -XOverloadedLabel instances unwrap all the way to the underlying numeric value.

>>> view #unNetBytes x1
7

>>> view #unSomeNetSize x2
7

>>> let x3 = hideDirection x1 :: SomeNetDir G Int
>>> view #unSomeNetDir x3
7

>>> let x4 = hideSize x3 :: SomeNet Int
>>> view #unSomeNet x4
7

Transformations

Converting Units

class Conversion a where

This class allows one to transform a bytes type to any Size. For types with existentially quantified Size (e.g. SomeSize, SomeNetSize), this will "undo" the existential quantification.

Since: 0.1

Associated Types

type Converted (t :: Size) a = (r :: Type) | r -> t

Since: 0.1

Methods

convert_ :: forall (t :: Size). SingI t => a -> Converted t a

convert_ @_ @t x converts x to size t.

Examples

Expand

>>> let bytes = MkBytes 50_000 :: Bytes 'M Int
>>> let gBytes = convert_ @_ @G bytes
>>> :type gBytes
gBytes :: Bytes G Int
>>> gBytes
MkBytes 50

>>> let bytes = hideSize (MkBytes 0.2 :: Bytes 'T Float)
>>> let mBytes = convert_ @_ @M bytes
>>> :type mBytes
mBytes :: Bytes M Float
>>> mBytes
MkBytes 200000.0

Since: 0.1

Normalization

class Normalize a where

Used for normalizing bytes b such that

\[ 1 \le \text{normalize}(b) < 1000 \iff 1\text{ B} \le b < 1000\text{ Y}. \]

In the strictest sense, \(\textrm{normalize} : \mathcal{C} \rightarrow \mathcal{C}\) is not a homomorphism, as the combination of two normalized values may itself not be normalized.

However, because the normalized units varies with the value, normalize always returns a type that existentially quantifies the size (e.g. SomeSize). Eq for these types is defined in terms of an equivalence class that takes units into account e.g. 1 P = 1,000 T = 1,000,000 G .... Viewed this way, normalize is actually an isomorphism, as it is essentially a no-op, never leaving the equivalence class.

This means we can happily mix normalization with different functions without worrying about the order. The only requirement we have is that such functions respect substitution:

\[ x = y \implies f(x) = f(y). \]

This is certainly true for all the usual mathematical operations we would normally use, e.g., AGroup addition, Module scalar multiplication. On the other hand, any functions that inspect the underlying numeric value or Bytes types could easily break this law. As such they should be treated with suspicion, at least when used in conjunction with normalize.

The other consideration we must keep in mind is that the final result of a series of computations may not be normalized. If this is desired, then normalize should be the last operation performed. Using normalize in the middle would not cause any harm (other than, perhaps, impacting efficiency), but it would not guarantee the final result is normalized.

Since: 0.1

Associated Types

type Norm a

Since: 0.1

Methods

normalize :: a -> Norm a

Normalizes the value.

Examples

Expand

>>> let bytes = MkBytes 5000 :: Bytes 'M Int
>>> normalize bytes
MkSomeSize SG (MkBytes 5)

>>> let bytes = hideSize (MkBytes 0.01 :: Bytes 'T Float)
>>> normalize bytes
MkSomeSize SG (MkBytes 10.0)

Since: 0.1

Algebra

The built-in Num class is abandoned in favor of algebra-simple's algebraic hierarchy based on abstract algebra. This is motivated by a desire to:

Provide a consistent API.

Avoid Num's infelicities (e.g. nonsense multiplication, dangerous fromInteger).

NetBytes and SomeNetSize are both AGroups. A Ring instance is not provided because multiplication is nonsensical:

\[ x \;\textrm{mb} \times y \;\textrm{mb} = xy \;\textrm{mb}^2. \]

Fortunately, multiplying bytes by some kind of scalar is both useful and has an easy interpretation: NetBytes forms a Module over a Ring (resp. VectorSpace over a Field). This allows us to multiply a NetBytes or SomeNetSize by a scalar in a manner consistent with the above API.

Examples

Addition/Subtraction

>>> import Numeric.Algebra (ASemigroup ((.+.)), AGroup ((.-.)))
>>> let mb1 = MkNetBytesP 20 :: NetBytes 'Down 'M Int
>>> let mb2 = MkNetBytesP 50 :: NetBytes 'Down 'M Int
>>> mb1 .+. mb2
MkNetBytes (MkBytes 70)
>>> mb1 .-. mb2
MkNetBytes (MkBytes (-30))

>>> let kb = MkNetBytesP 50 :: NetBytes 'Down 'K Int
>>> -- mb1 .+. kb -- This would be a type error

>>> let mbUp = MkNetBytesP 50 :: NetBytes 'Up 'M Int
>>> -- mb1 .+. mbUp -- This would be a type error

Multiplication

>>> import Numeric.Algebra (MSemiSpace ((.*)))
>>> mb1 .* 10
MkNetBytes (MkBytes 200)

Division

>>> import Numeric.Algebra (MSpace ((.%)))
>>> mb1 .% 10
MkNetBytes (MkBytes 2)

One may wonder how the AGroup instance for SomeNetSize could possibly work. It is possible (indeed, expected) that we could have two SomeNetSizes that have different underlying NetBytes types. To handle this, the SomeNetSize instance will convert both NetBytes to a NetBytes 'B before adding/subtracting.

>>> let some1 = hideSize (MkNetBytesP 1000 :: NetBytes 'Down 'G Double)
>>> let some2 = hideSize (MkNetBytesP 500_000 :: NetBytes 'Down 'M Double)
>>> some1 .+. some2
MkSomeNetSize SB (MkNetBytes (MkBytes 1.5e12))
>>> some1 .-. some2
MkSomeNetSize SB (MkNetBytes (MkBytes 5.0e11))

This respects SomeNetSize's equivalence-class base Eq.

module Numeric.Algebra

module Numeric.Convert.Integer

module Numeric.Convert.Rational

module Numeric.Convert.Real

Text

Pretty Printing

We provide several formatters for pretty-printing different byte types.

>>> import Data.Default (Default (def))
>>> let bf = MkFloatingFormatter (Just 2)
>>> let b = MkNetBytesP @Up @G @Float 203.301
>>> formatSizedDirected bf def def b
"203.30 gb up"

module Data.Bytes.Formatting

Parsing

class Parser a

Represents a megaparsec parser. Used for parsing byte types from Text.

Since: 0.1

Minimal complete definition

parser

parse :: Parser a => Text -> Either Text a

Parses various byte types from Text. Parsing is lenient in general. We support:

Case-insensitivity.

Optional leading/internal/trailing whitespace.

Flexible names.

Bytes Examples

Expand

>>> import Data.Bytes (Bytes, Size (..), SomeSize)
>>> parse @(Bytes M Int) "70"
Right (MkBytes 70)

>>> parse @(SomeSize Float) "100.45 kilobytes"
Right (MkSomeSize SK (MkBytes 100.45))

>>> parse @(SomeSize Word) "2300G"
Right (MkSomeSize SG (MkBytes 2300))

>>> parse @(SomeSize Float) "5.5 tb"
Right (MkSomeSize ST (MkBytes 5.5))

Network Examples

Expand

>>> import Data.Bytes.Network (Direction (..), NetBytes, SomeNet, SomeNetDir, SomeNetSize)
>>> parse @(NetBytes Up M Int) "70"
Right (MkNetBytes (MkBytes 70))

>>> parse @(SomeNetSize Down Float) "100.45 kilobytes"
Right (MkSomeNetSize SK (MkNetBytes (MkBytes 100.45)))

>>> parse @(SomeNetSize Up Word) "2300G"
Right (MkSomeNetSize SG (MkNetBytes (MkBytes 2300)))

>>> parse @(SomeNetDir T Word) "2300 up"
Right (MkSomeNetDir SUp (MkNetBytes (MkBytes 2300)))

>>> parse @(SomeNetDir M Word) "2300D"
Right (MkSomeNetDir SDown (MkNetBytes (MkBytes 2300)))

>>> parse @(SomeNet Float) "5.5 tb Up"
Right (MkSomeNet SUp ST (MkNetBytes (MkBytes 5.5)))

>>> parse @(SomeNet Float) "5.5 megabytes DOWN"
Right (MkSomeNet SDown SM (MkNetBytes (MkBytes 5.5)))

Since: 0.1

Optics

Core

_MkNetBytes :: forall (s :: Size) (d :: Direction) n. Iso' (NetBytes d s n) (Bytes s n)

Iso' between NetBytes and underlying Bytes.

Examples

Expand

>>> import Optics.Core (review, view)
>>> review _MkNetBytes (MkBytes @K @Int 70)
MkNetBytes (MkBytes 70)

>>> view _MkNetBytes (MkNetBytes $ MkBytes @K @Int 70)
MkBytes 70

Since: 0.1

_MkSomeNetSize :: forall (s :: Size) (d :: Direction) n. (FromInteger n, MGroup n, SingI s) => Iso' (SomeNetSize d n) (NetBytes d s n)

Iso' between SomeNetSize and underlying NetBytes. Performs any necessary conversions when going from SomeNetSize d n -> NetBytes d s n.

Examples

Expand

>>> import Optics.Core (review, view)
>>> review _MkSomeNetSize (MkNetBytesP @Up @K @Int 70)
MkSomeNetSize SK (MkNetBytes (MkBytes 70))

>>> (view _MkSomeNetSize (hideSize $ MkNetBytesP @Up @K @Int 70)) :: NetBytes Up B Int
MkNetBytes (MkBytes 70000)

Since: 0.1

_MkSomeNetDir :: forall (s :: Size) (d :: Direction) n. SingI d => Review (SomeNetDir s n) (NetBytes d s n)

Review between SomeNetDir and underlying NetBytes. This is not an iso (i.e. only allows NetBytes -> SomeNetDir) because the opposite direction would require dropping the Direction and the user arbitrarily choosing a new one.

Examples

Expand

>>> import Optics.Core (review)
>>> review _MkSomeNetDir (MkNetBytesP @Up @K @Int 70)
MkSomeNetDir SUp (MkNetBytes (MkBytes 70))

Since: 0.1

_MkSomeNet :: forall (s :: Size) (d :: Direction) n. (SingI d, SingI s) => Review (SomeNet n) (NetBytes d s n)

Review between SomeNet and underlying NetBytes. This is not an iso (i.e. only allows NetBytes -> SomeNet) because the opposite direction would require dropping the Direction and the user arbitrarily choosing a new one.

Examples

Expand

>>> import Optics.Core (review)
>>> review _MkSomeNet (MkNetBytesP @Up @K @Int 70)
MkSomeNet SUp SK (MkNetBytes (MkBytes 70))

Since: 0.1

Size

_B :: Prism' Size ()

Since: 0.1

_K :: Prism' Size ()

Since: 0.1

_M :: Prism' Size ()

Since: 0.1

_G :: Prism' Size ()

Since: 0.1

_T :: Prism' Size ()

Since: 0.1

_P :: Prism' Size ()

Since: 0.1

_E :: Prism' Size ()

Since: 0.1

_Z :: Prism' Size ()

Since: 0.1

_Y :: Prism' Size ()

Since: 0.1

Direction

_Down :: Prism' Direction ()

Since: 0.1

_Up :: Prism' Direction ()

Since: 0.1

Reexports

class Default a where

A class for types with a default value.

Minimal complete definition

Nothing

Methods

def :: a

The default value for this type.

default def :: (Generic a, GDefault (Rep a)) => a