Data.Bytes

Description

This module serves as the main entry point for the library. It provides the types and operations for typical usage and is usually the only import required. The core concept is:

Wrapping a numeric value representing bytes in a new type.

Attaching phantom labels representing the units (e.g. K, M, ...).

This prevents mistakes, such as adding two different byte sizes or converting between sizes incorrectly.

Since: 0.1

Introduction

The main idea is to attach phantom labels to the numeric bytes, so we can track the size units. This allows us to safely manipulate byte values without mixing up units, performing incorrect conversions, etc.

The core types are a newtype wrapper Bytes and the Size units:

newtype Bytes (s :: Size) n

This is the core type for handling type-safe byte operations. It is intended to be used as a simple wrapper over some numeric type, equipped with a Size tag.

To take full advantage of the API (e.g. normalize), the underlying numeric type should implement Semifield or, ideally, Field.

Examples

Expand

>>> MkBytes @M 1000
MkBytes 1000

Since: 0.1

Constructors

MkBytes n

data Size

Byte units.

Since: 0.1

Constructors

B
Bytes

Since: 0.1

K
Kilobytes

Since: 0.1

M
Megabytes

Since: 0.1

G
Gigabytes

Since: 0.1

T
Terabytes

Since: 0.1

P
Petabytes

Since: 0.1

E
Exabytes

Since: 0.1

Z
Zettabytes

Since: 0.1

Y
Yottabytes

Since: 0.1

Basic Usage

Construction

There are several ways to construct a Bytes type.

FromInteger

>>> fromZ 80 :: Bytes M Int
MkBytes 80

Directly

>>> MkBytes 80 :: Bytes M Int
MkBytes 80

Optics (optics-core)

>>> import Optics.Core (review)
>>> (review _MkBytes 70) :: Bytes G Int
MkBytes 70

>>> (review #unBytes 70) :: Bytes G Int
MkBytes 70

Unknown Size

We sometimes have to deal with unknown sizes at runtime, which presents a problem. We handle this with the SomeSize type, which existentially quantifies the Size:

data SomeSize n

Wrapper for Bytes, existentially quantifying the size. This is useful when a function does not know a priori what size it should return e.g.

>>> :{
  getFileSize :: FilePath -> IO (SomeSize Float)
  getFileSize path = do
    -- getRawFileSize :: FilePath -> IO (Float, String)
    (bytes, units) <- getRawFileSize path
    pure $ case units of
      "B" -> hideSize $ MkBytes @B bytes
      "K" -> hideSize $ MkBytes @K bytes
      _ -> error "todo"
:}

We define an equivalence relation on SomeSize that takes units into account. For instance,

>>> hideSize (MkBytes @G 7) == hideSize (MkBytes @M 7_000)
True

Because we expose the underlying Bytes in several ways (e.g. Show, the SSize witness), this is technically unlawful for equality as it breaks the extensionality law:

\[ x = y \implies f(x) = f(y). \]

Since: 0.1

Sized

Fortunately, we do not have to directly use the constructor or singletons. We can instead use the Sized class.

class Sized a where

Types that have a size.

Since: 0.1

Associated Types

type HideSize a

Type used to hide the size.

Since: 0.1

Methods

sizeOf :: a -> Size

Retrieves the size.

Examples

Expand

>>> import Data.Bytes (Bytes (..))
>>> sizeOf (MkBytes @G 7)
G

>>> sizeOf (hideSize $ MkBytes @M 7)
M

>>> import Data.Bytes.Network (NetBytes (..), Direction (..))
>>> sizeOf (hideSize $ MkNetBytesP @Up @M 7)
M

Since: 0.1

hideSize :: a -> HideSize a

Hides the size.

Examples

Expand

>>> import Data.Bytes (Bytes (..))
>>> hideSize (MkBytes @G 7)
MkSomeSize SG (MkBytes 7)

Optics

Once again, we can use optics for this.

>>> import Optics.Core (review)
>>> review _MkSomeSize (MkBytes 70 :: Bytes G Int)
MkSomeSize SG (MkBytes 70)

Elimination

RawNumeric

We provide the RawNumeric class for conveniently unwrapping a type to the underlying numeric value.

class RawNumeric a where

Abstracts "wrapper" types for generically retrieving a raw numeric value.

Since: 0.1

Associated Types

type Raw a

Since: 0.1

Methods

toRaw :: a -> Raw a

Retrieves the underlying value.

Examples

Expand

>>> import Data.Bytes (Bytes (..), Size (..), Sized (..))
>>> toRaw (MkBytes @G 7)
7

>>> toRaw (hideSize $ MkBytes @M 400)
400

>>> import Data.Bytes.Network (Direction (..), NetBytes (..))
>>> toRaw (MkNetBytesP @Up @G 7)
7

>>> toRaw (hideSize $ MkNetBytesP @Up @G 7)
7

Since: 0.1

HasField

We can use HasField for this too.

>>> -- {-# LANGUAGE OverloadedRecordDot #-}
>>> let x = MkBytes 7 :: Bytes G Int
>>> x.unBytes
7

>>> let y = hideSize x :: SomeSize Int
>>> y.unSomeSize
7

Optics

Optics are another option. The underscore-prefixed optics unwrap one level at a time, since we can freely compose them.

>>> import Optics.Core (view, (%))
>>> let x = MkBytes 7 :: Bytes G Int
>>> view _MkBytes x
7

>>> -- notice we have to convert the numeric value since the requested
>>> -- return type ('M') differs from the original ('G')
>>> let y = hideSize x :: SomeSize Int
>>> (view _MkSomeSize y) :: Bytes M Int
MkBytes 7000

>>> view (_MkSomeSize % (_MkBytes @M)) y
7000

The -XOverloadedLabel instances unwrap all the way to the underlying numeric value.

>>> view #unBytes x
7

>>> view #unSomeSize y
7

Transformations

Converting Units

class Conversion a where

This class allows one to transform a bytes type to any Size. For types with existentially quantified Size (e.g. SomeSize, SomeNetSize), this will "undo" the existential quantification.

Since: 0.1

Associated Types

type Converted (t :: Size) a = (r :: Type) | r -> t

Since: 0.1

Methods

convert_ :: forall (t :: Size). SingI t => a -> Converted t a

convert_ @_ @t x converts x to size t.

Examples

Expand

>>> let bytes = MkBytes 50_000 :: Bytes 'M Int
>>> let gBytes = convert_ @_ @G bytes
>>> :type gBytes
gBytes :: Bytes G Int
>>> gBytes
MkBytes 50

>>> let bytes = hideSize (MkBytes 0.2 :: Bytes 'T Float)
>>> let mBytes = convert_ @_ @M bytes
>>> :type mBytes
mBytes :: Bytes M Float
>>> mBytes
MkBytes 200000.0

Since: 0.1

Normalization

class Normalize a where

Used for normalizing bytes b such that

\[ 1 \le \text{normalize}(b) < 1000 \iff 1\text{ B} \le b < 1000\text{ Y}. \]

In the strictest sense, \(\textrm{normalize} : \mathcal{C} \rightarrow \mathcal{C}\) is not a homomorphism, as the combination of two normalized values may itself not be normalized.

However, because the normalized units varies with the value, normalize always returns a type that existentially quantifies the size (e.g. SomeSize). Eq for these types is defined in terms of an equivalence class that takes units into account e.g. 1 P = 1,000 T = 1,000,000 G .... Viewed this way, normalize is actually an isomorphism, as it is essentially a no-op, never leaving the equivalence class.

This means we can happily mix normalization with different functions without worrying about the order. The only requirement we have is that such functions respect substitution:

\[ x = y \implies f(x) = f(y). \]

This is certainly true for all the usual mathematical operations we would normally use, e.g., AGroup addition, Module scalar multiplication. On the other hand, any functions that inspect the underlying numeric value or Bytes types could easily break this law. As such they should be treated with suspicion, at least when used in conjunction with normalize.

The other consideration we must keep in mind is that the final result of a series of computations may not be normalized. If this is desired, then normalize should be the last operation performed. Using normalize in the middle would not cause any harm (other than, perhaps, impacting efficiency), but it would not guarantee the final result is normalized.

Since: 0.1

Associated Types

type Norm a

Since: 0.1

Methods

normalize :: a -> Norm a

Normalizes the value.

Examples

Expand

>>> let bytes = MkBytes 5000 :: Bytes 'M Int
>>> normalize bytes
MkSomeSize SG (MkBytes 5)

>>> let bytes = hideSize (MkBytes 0.01 :: Bytes 'T Float)
>>> normalize bytes
MkSomeSize SG (MkBytes 10.0)

Since: 0.1

Algebra

The built-in Num class is abandoned in favor of algebra-simple's algebraic hierarchy based on abstract algebra. This is motivated by a desire to:

Provide a consistent API.

Avoid Num's infelicities (e.g. nonsense multiplication, dangerous fromInteger).

Bytes and SomeSize are both AGroups. A Ring instance is not provided because multiplication is nonsensical:

\[ x \;\textrm{mb} \times y \;\textrm{mb} = xy \;\textrm{mb}^2. \]

Fortunately, multiplying bytes by some kind of scalar is both useful and has an easy interpretation: Bytes forms a Module over a Ring (resp. VectorSpace over a Field). This allows us to multiply a Bytes or SomeSize by a scalar in a manner consistent with the above API.

Examples

Addition/Subtraction

>>> import Numeric.Algebra (ASemigroup ((.+.)), AGroup ((.-.)))
>>> let mb1 = MkBytes 20 :: Bytes 'M Int
>>> let mb2 = MkBytes 50 :: Bytes 'M Int
>>> mb1 .+. mb2
MkBytes 70
>>> mb1 .-. mb2
MkBytes (-30)

>>> let kb = MkBytes 50 :: Bytes 'K Int
>>> -- mb1 .+. kb -- This would be a type error

Multiplication

>>> import Numeric.Algebra (MSemiSpace ((.*)))
>>> mb1 .* 10
MkBytes 200

Division

>>> import Numeric.Algebra (MSpace ((.%)))
>>> mb1 .% 10
MkBytes 2

One may wonder how the AGroup instance for SomeSize could possibly work. It is possible (indeed, expected) that we could have two SomeSizes that have different underlying Bytes types. To handle this, the SomeSize instance will convert both Bytes to a Bytes 'B before adding/subtracting.

>>> let some1 = hideSize (MkBytes 1000 :: Bytes 'G Double)
>>> let some2 = hideSize (MkBytes 500_000 :: Bytes 'M Double)
>>> some1 .+. some2
MkSomeSize SB (MkBytes 1.5e12)
>>> some1 .-. some2
MkSomeSize SB (MkBytes 5.0e11)

This respects SomeSize's equivalence-class based Eq.

module Numeric.Algebra

module Numeric.Convert.Integer

module Numeric.Convert.Rational

module Numeric.Convert.Real

Text

Pretty Printing

We provide several formatters for pretty-printing different byte types.

>>> import Data.Default (Default (def))
>>> let bf = MkFloatingFormatter (Just 2)
>>> let b = MkBytes @G @Float 20.248
>>> formatSized bf def b
"20.25 gb"

module Data.Bytes.Formatting

Parsing

class Parser a

Represents a megaparsec parser. Used for parsing byte types from Text.

Since: 0.1

Minimal complete definition

parser

parse :: Parser a => Text -> Either Text a

Parses various byte types from Text. Parsing is lenient in general. We support:

Case-insensitivity.

Optional leading/internal/trailing whitespace.

Flexible names.

Bytes Examples

Expand

>>> import Data.Bytes (Bytes, Size (..), SomeSize)
>>> parse @(Bytes M Int) "70"
Right (MkBytes 70)

>>> parse @(SomeSize Float) "100.45 kilobytes"
Right (MkSomeSize SK (MkBytes 100.45))

>>> parse @(SomeSize Word) "2300G"
Right (MkSomeSize SG (MkBytes 2300))

>>> parse @(SomeSize Float) "5.5 tb"
Right (MkSomeSize ST (MkBytes 5.5))

Network Examples

Expand

>>> import Data.Bytes.Network (Direction (..), NetBytes, SomeNet, SomeNetDir, SomeNetSize)
>>> parse @(NetBytes Up M Int) "70"
Right (MkNetBytes (MkBytes 70))

>>> parse @(SomeNetSize Down Float) "100.45 kilobytes"
Right (MkSomeNetSize SK (MkNetBytes (MkBytes 100.45)))

>>> parse @(SomeNetSize Up Word) "2300G"
Right (MkSomeNetSize SG (MkNetBytes (MkBytes 2300)))

>>> parse @(SomeNetDir T Word) "2300 up"
Right (MkSomeNetDir SUp (MkNetBytes (MkBytes 2300)))

>>> parse @(SomeNetDir M Word) "2300D"
Right (MkSomeNetDir SDown (MkNetBytes (MkBytes 2300)))

>>> parse @(SomeNet Float) "5.5 tb Up"
Right (MkSomeNet SUp ST (MkNetBytes (MkBytes 5.5)))

>>> parse @(SomeNet Float) "5.5 megabytes DOWN"
Right (MkSomeNet SDown SM (MkNetBytes (MkBytes 5.5)))

Since: 0.1

Optics

Core

_MkBytes :: forall (s :: Size) n. Iso' (Bytes s n) n

Iso' between Bytes and underlying value.

Examples

Expand

>>> import Optics.Core (review, view)
>>> (review _MkBytes 70) :: Bytes K Int
MkBytes 70

>>> view _MkBytes (MkBytes @K @Int 70)
70

Since: 0.1

_MkSomeSize :: forall (s :: Size) n. (FromInteger n, MGroup n, SingI s) => Iso' (SomeSize n) (Bytes s n)

Iso' between SomeSize and underlying Bytes. Performs any necessary conversions when going from SomeSize n -> Bytes s n.

Examples

Expand

>>> import Optics.Core (review, view)
>>> review _MkSomeSize (MkBytes @K @Int 70)
MkSomeSize SK (MkBytes 70)

>>> (view _MkSomeSize (hideSize $ MkBytes @K @Int 70)) :: Bytes B Int
MkBytes 70000

Since: 0.1

Size

_B :: Prism' Size ()

Since: 0.1

_K :: Prism' Size ()

Since: 0.1

_M :: Prism' Size ()

Since: 0.1

_G :: Prism' Size ()

Since: 0.1

_T :: Prism' Size ()

Since: 0.1

_P :: Prism' Size ()

Since: 0.1

_E :: Prism' Size ()

Since: 0.1

_Z :: Prism' Size ()

Since: 0.1

_Y :: Prism' Size ()

Since: 0.1

Reexports

class Default a where

A class for types with a default value.

Minimal complete definition

Nothing

Methods

def :: a

The default value for this type.

default def :: (Generic a, GDefault (Rep a)) => a