Data.Bytes.Network.Internal

Description

Internal module for Data.Network.NetBytes. The primary difference is this module exposes some underlying details i.e. singleton witnesses. These are hidden by default as they complicate the API.

Since: 0.1

Network Bytes

newtype NetBytes (d :: Direction) (s :: Size) n

Wrapper around the Bytes type that adds the Direction tag.

Examples

Expand

>>> MkNetBytesP @Up @M 1000
MkNetBytes (MkBytes 1000)

Since: 0.1

Constructors

MkNetBytes (Bytes s n)

Bundled Patterns

pattern MkNetBytesP :: forall d s n. n -> NetBytes d s n
Pattern for de/constructing NetBytes.

Since: 0.1

_MkNetBytes :: forall (s :: Size) (d :: Direction) n. Iso' (NetBytes d s n) (Bytes s n)

Iso' between NetBytes and underlying Bytes.

Examples

Expand

>>> import Optics.Core (review, view)
>>> review _MkNetBytes (MkBytes @K @Int 70)
MkNetBytes (MkBytes 70)

>>> view _MkNetBytes (MkNetBytes $ MkBytes @K @Int 70)
MkBytes 70

Since: 0.1

netToSSize :: forall (s :: Size) (d :: Direction) n. SingI s => NetBytes d s n -> SSize s

Retrieves the SingI witness. Can be used to recover the Size.

>>> netToSSize (MkNetBytesP @Down @K @Int 7)
SK

Since: 0.1

netToSDirection :: forall (d :: Direction) (s :: Size) n. SingI d => NetBytes d s n -> SDirection d

Retrieves the SDirection witness. Can be used to recover the Direction.

>>> netToSDirection (MkNetBytesP @Up @K @Int 7)
SUp

Since: 0.1

Unknown Size

data SomeNetSize (d :: Direction) n where

Wrapper for NetBytes, existentially quantifying the size. This is useful when a function does not know a priori what size it should return e.g.

>>> :{
  getUpTraffic :: IO (SomeNetSize Up Float)
  getUpTraffic = do
    -- getUpTrafficRaw :: IO (Float, String)
    (bytes, units) <- getUpTrafficRaw
    pure $ case units of
      "B" -> hideSize $ MkNetBytesP @Up @B bytes
      "K" -> hideSize $ MkNetBytesP @Up @K bytes
      _ -> error "todo"
:}

We define an equivalence relation on SomeNetSize that takes units into account. For instance,

>>> hideSize (MkNetBytesP @Up @K 1000) == hideSize (MkNetBytesP @Up @M 1)
True

Because we expose the underlying NetBytes in several ways (e.g. Show, the SSize witness), this is technically unlawful for equality as it breaks the extensionality law:

\[ x = y \implies f(x) = f(y). \]

Since: 0.1

Constructors

MkSomeNetSize :: forall (s :: Size) (d :: Direction) n. SSize s -> NetBytes d s n -> SomeNetSize d n
Since: 0.1

_MkSomeNetSize :: forall (s :: Size) (d :: Direction) n. (FromInteger n, MGroup n, SingI s) => Iso' (SomeNetSize d n) (NetBytes d s n)

Iso' between SomeNetSize and underlying NetBytes. Performs any necessary conversions when going from SomeNetSize d n -> NetBytes d s n.

Examples

Expand

>>> import Optics.Core (review, view)
>>> review _MkSomeNetSize (MkNetBytesP @Up @K @Int 70)
MkSomeNetSize SK (MkNetBytes (MkBytes 70))

>>> (view _MkSomeNetSize (hideSize $ MkNetBytesP @Up @K @Int 70)) :: NetBytes Up B Int
MkNetBytes (MkBytes 70000)

Since: 0.1

someNetSizeToSDirection :: forall (d :: Direction) n. SingI d => SomeNetSize d n -> SDirection d

Retrieves the SingI witness. Can be used to recover the Direction.

Since: 0.1

Unknown Direction

data SomeNetDir (s :: Size) n where

Wrapper for NetBytes, existentially quantifying the direction. This is useful when a function does not know a priori what direction it should return e.g.

>>> :{
  getMaxTraffic :: IO (SomeNetDir K Double)
  getMaxTraffic = do
    -- getMaxTraffickRaw :: IO (Double, String)
    (bytes, direction) <- getMaxTraffickRaw
    pure $ case direction of
      "down" -> hideDirection $ MkNetBytesP @Down bytes
      "up" -> hideDirection $ MkNetBytesP @Up bytes
      _ -> error "bad direction"
:}

We deliberately do not provide instances for SomeX classes that could be used to combine arbitrary SomeNetDirs (e.g. Applicative, AGroup), as that would defeat the purpose of enforcing the distinction between upload and downloaded bytes.

Equality is determined by the usual equivalence class -- that takes units into account -- and by considering the direction.

>>> let x = MkNetBytesP @Up @K 1000 :: NetBytes Up K Int
>>> let y = MkNetBytesP @Down @K 1000 :: NetBytes Down K Int
>>> hideDirection x == hideDirection x
True
>>> hideDirection x == hideDirection y
False

Notice no Ord instance is provided, as we provide no ordering for Direction.

Since: 0.1

Constructors

MkSomeNetDir :: forall (d :: Direction) (s :: Size) n. SDirection d -> NetBytes d s n -> SomeNetDir s n
Since: 0.1

_MkSomeNetDir :: forall (s :: Size) (d :: Direction) n. SingI d => Review (SomeNetDir s n) (NetBytes d s n)

Review between SomeNetDir and underlying NetBytes. This is not an iso (i.e. only allows NetBytes -> SomeNetDir) because the opposite direction would require dropping the Direction and the user arbitrarily choosing a new one.

Examples

Expand

>>> import Optics.Core (review)
>>> review _MkSomeNetDir (MkNetBytesP @Up @K @Int 70)
MkSomeNetDir SUp (MkNetBytes (MkBytes 70))

Since: 0.1

someNetDirToSSize :: forall (s :: Size) n. SingI s => SomeNetDir s n -> SSize s

Retrieves the SingI witness. Can be used to recover the Size.

Since: 0.1

Unknown Direction and Size

data SomeNet n where

Wrapper for NetBytes, existentially quantifying the size and direction. This is useful when a function does not know a priori what size or direction it should return e.g.

>>> :{
  getMaxTraffic :: IO (SomeNet Double)
  getMaxTraffic = do
    -- getMaxTrafficNetRaw :: IO (Double, String, String)
    (bytes, direction, size) <- getMaxTrafficNetRaw
    pure $ case (direction, size) of
      ("Down", "K") -> hideDirection $ hideSize $ MkNetBytesP @Down @K bytes
      ("Up", "M") -> hideDirection $ hideSize $ MkNetBytesP @Up @M bytes
      _ -> error "todo"
:}

SomeNet carries along SDirection and SSize runtime witnesses for recovering the Direction and Size, respectively.

This instance uses the same equivalence relation from SomeNetSize w.r.t the size, and also includes an equality check on the direction. Thus we have, for instance,

>>> let x = MkNetBytesP 1_000 :: NetBytes Up K Int
>>> let y = MkNetBytesP 1 :: NetBytes Up M Int
>>> let z = MkNetBytesP 1_000 :: NetBytes Down K Int
>>> hideDirection (hideSize x) == hideDirection (hideSize y)
True

>>> hideDirection (hideSize x) == hideDirection (hideSize z)
False

Since: 0.1

Constructors

MkSomeNet :: forall (d :: Direction) (s :: Size) n. SDirection d -> SSize s -> NetBytes d s n -> SomeNet n
Since: 0.1

_MkSomeNet :: forall (s :: Size) (d :: Direction) n. (SingI d, SingI s) => Review (SomeNet n) (NetBytes d s n)

Review between SomeNet and underlying NetBytes. This is not an iso (i.e. only allows NetBytes -> SomeNet) because the opposite direction would require dropping the Direction and the user arbitrarily choosing a new one.

Examples

Expand

>>> import Optics.Core (review)
>>> review _MkSomeNet (MkNetBytesP @Up @K @Int 70)
MkSomeNet SUp SK (MkNetBytes (MkBytes 70))

Since: 0.1